[integ-tests] Add integ test for DFSM
