+++ tests/integration-tests/tests/storage/test_fsx_lustre.py
@@ -18,20 +18,17 @@ import pytest
 import utils
 from assertpy import assert_that
 from botocore.exceptions import ClientError
-from cfn_stacks_factory import CfnStack
 from remote_command_executor import RemoteCommandExecutor
 from retrying import retry
 from time_utils import minutes, seconds
-from troposphere import Template
-from troposphere.fsx import (
-    ClientConfigurations,
-    LustreConfiguration,
-    NfsExports,
-    OntapConfiguration,
-    OpenZFSConfiguration,
-    RootVolumeConfiguration,
-    Volume,
-    VolumeOpenZFSConfiguration,
+from troposphere.fsx import LustreConfiguration
+
+from tests.storage.storage_common import (
+    assert_fsx_lustre_correctly_mounted,
+    check_fsx,
+    create_fsx_ontap,
+    create_fsx_open_zfs,
+    get_fsx_ids,
 )
 
 BACKUP_NOT_YET_AVAILABLE_STATES = {"CREATING", "TRANSFERRING", "PENDING"}
@@ -180,34 +177,6 @@ def test_fsx_lustre(
     )
 
 
-def check_fsx(
-    cluster,
-    region,
-    scheduler_commands_factory,
-    mount_dirs,
-    bucket_name,
-):
-    remote_command_executor = RemoteCommandExecutor(cluster)
-    scheduler_commands = scheduler_commands_factory(remote_command_executor)
-    fsx_ids = get_fsx_ids(cluster, region)
-    logging.info("Checking the length of mount dirs is the same as the length of FSXIDs")
-    assert_that(len(mount_dirs)).is_equal_to(len(fsx_ids))
-    for mount_dir, fsx_id in zip(mount_dirs, fsx_ids):
-        logging.info("Checking %s on %s", fsx_id, mount_dir)
-        file_system_type = get_file_system_type(fsx_id, region)
-        if file_system_type == "LUSTRE":
-            assert_fsx_lustre_correctly_mounted(remote_command_executor, mount_dir, region, fsx_id)
-            if bucket_name:
-                _test_import_path(remote_command_executor, mount_dir)
-                _test_export_path(remote_command_executor, mount_dir, bucket_name, region)
-                _test_data_repository_task(remote_command_executor, mount_dir, bucket_name, fsx_id, region)
-        elif file_system_type == "OPENZFS":
-            assert_fsx_open_zfs_correctly_mounted(remote_command_executor, mount_dir, fsx_id)
-        elif file_system_type == "ONTAP":
-            assert_fsx_ontap_correctly_mounted(remote_command_executor, mount_dir, fsx_id)
-        assert_fsx_correctly_shared(scheduler_commands, remote_command_executor, mount_dir)
-
-
 @pytest.mark.usefixtures("os", "instance", "scheduler")
 def test_fsx_lustre_backup(region, pcluster_config_reader, clusters_factory, scheduler_commands_factory):
     """
@@ -273,93 +242,6 @@ def test_fsx_lustre_backup(region, pcluster_config_reader, clusters_factory, sch
 
     # Test deletion of manual backup
     _test_delete_manual_backup(manual_backup, region)
-
-
-def create_fsx_ontap(fsx_factory, num):
-    return fsx_factory(
-        ports=[111, 635, 2049, 4046],
-        ip_protocols=["tcp", "udp"],
-        num=num,
-        file_system_type="ONTAP",
-        StorageCapacity=1024,
-        OntapConfiguration=OntapConfiguration(DeploymentType="SINGLE_AZ_1", ThroughputCapacity=128),
-    )
-
-
-def create_fsx_open_zfs(fsx_factory, num):
-    if num == 0:
-        return []
-    file_system_ids = fsx_factory(
-        ports=[111, 2049, 20001, 20002, 20003],
-        ip_protocols=["tcp", "udp"],
-        num=num,
-        file_system_type="OPENZFS",
-        StorageCapacity=64,
-        OpenZFSConfiguration=OpenZFSConfiguration(
-            DeploymentType="SINGLE_AZ_1",
-            ThroughputCapacity=64,
-            RootVolumeConfiguration=RootVolumeConfiguration(
-                NfsExports=[
-                    NfsExports(ClientConfigurations=[ClientConfigurations(Clients="*", Options=["rw", "crossmnt"])])
-                ]
-            ),
-        ),
-    )
-    volume_list = boto3.client("fsx").describe_volumes(Filters=[{"Name": "file-system-id", "Values": file_system_ids}])[
-        "Volumes"
-    ]
-    return [volume["VolumeId"] for volume in volume_list]
-
-
-@pytest.fixture(scope="class")
-def open_zfs_volume_factory(vpc_stack, cfn_stacks_factory, request, region, key_name):
-    """
-    Define a fixture to manage the creation and destruction of volumes for FSx for OpenZFS.
-
-    return volume ids
-    """
-    created_stacks = []
-
-    def _open_zfs_volume_factory(root_volume_id, num_volumes=1):
-        fsx_open_zfs_volume_template = Template()
-        fsx_open_zfs_volume_template.set_version()
-        fsx_open_zfs_volume_template.set_description("Create Storage Virtual Machine stack")
-
-        open_zfs_volume_resource_name = "OpenZFSVolume"
-        max_concurrency = 15
-        for index in range(num_volumes):
-            depends_on_arg = {}
-            if index >= max_concurrency:
-                depends_on_arg = {"DependsOn": [f"{open_zfs_volume_resource_name}{index - max_concurrency}"]}
-            fsx_open_zfs_volume = Volume(
-                title=f"{open_zfs_volume_resource_name}{index}",
-                Name=f"vol{index}",
-                VolumeType="OPENZFS",
-                OpenZFSConfiguration=VolumeOpenZFSConfiguration(
-                    NfsExports=[
-                        NfsExports(ClientConfigurations=[ClientConfigurations(Clients="*", Options=["rw", "crossmnt"])])
-                    ],
-                    ParentVolumeId=root_volume_id,
-                ),
-                **depends_on_arg,
-            )
-            fsx_open_zfs_volume_template.add_resource(fsx_open_zfs_volume)
-        fsx_stack = CfnStack(
-            name=utils.generate_stack_name(
-                "integ-tests-fsx-openzfs-volume", request.config.getoption("stackname_suffix")
-            ),
-            region=region,
-            template=fsx_open_zfs_volume_template.to_json(),
-        )
-        cfn_stacks_factory.create_stack(fsx_stack)
-        created_stacks.append(fsx_stack)
-        return [fsx_stack.cfn_resources[f"{open_zfs_volume_resource_name}{i}"] for i in range(num_volumes)]
-
-    yield _open_zfs_volume_factory
-
-    if not request.config.getoption("no_delete"):
-        for stack in created_stacks:
-            cfn_stacks_factory.delete_stack(stack.name, region)
 
 
 @pytest.mark.usefixtures("instance", "scheduler")
@@ -460,86 +342,6 @@ def test_multiple_fsx(
     run_benchmarks(remote_command_executor, scheduler_commands)
 
 
-def assert_fsx_lustre_correctly_mounted(remote_command_executor, mount_dir, region, fsx_fs_id):
-    logging.info("Testing fsx lustre is correctly mounted on the head node")
-    result = remote_command_executor.run_remote_command("df -h -t lustre | tail -n +2 | awk '{print $1, $2, $6}'")
-    mount_name = get_mount_name(fsx_fs_id, region)
-    assert_that(result.stdout).matches(
-        r"[0-9\.]+@tcp:/{mount_name}\s+[15]\.[1278]T\s+{mount_dir}".format(mount_name=mount_name, mount_dir=mount_dir)
-    )
-    # example output: "192.168.46.168@tcp:/cg7k7bmv 1.7T /fsx_mount_dir"
-
-    mount_options = "defaults,_netdev,flock,user_xattr,noatime,noauto,x-systemd.automount"
-
-    check_fstab_file(
-        remote_command_executor,
-        r"fs-[0-9a-z]+\.fsx\.[a-z1-9\-]+\.amazonaws\.com@tcp:/{mount_name}"
-        r" {mount_dir} lustre {mount_options} 0 0".format(
-            mount_name=mount_name, mount_dir=mount_dir, mount_options=mount_options
-        ),
-    )
-
-
-def assert_fsx_open_zfs_correctly_mounted(remote_command_executor, mount_dir, volume_id):
-    logging.info("Testing fsx OpenZFS is correctly mounted on the head node")
-    result = remote_command_executor.run_remote_command("df -h -t nfs4")
-    fsx_client = boto3.client("fsx")
-    volume = fsx_client.describe_volumes(VolumeIds=[volume_id]).get("Volumes")[0]
-    volume_path = volume["OpenZFSConfiguration"]["VolumePath"]
-    fs_id = volume["FileSystemId"]
-    dns_name = fsx_client.describe_file_systems(FileSystemIds=[fs_id])["FileSystems"][0]["DNSName"]
-    remote_path = f"{dns_name}:{volume_path}"
-    assert_that(result.stdout).matches(rf"{remote_path} .* {mount_dir}")
-    # example output: "fs-123456.fsx.us-west-2.amazonaws.com:/fsx 64G 256K 64G 1% /fsx_mount_dir0"""
-    check_fstab_file(remote_command_executor, f"{remote_path} {mount_dir} nfs nfsvers=4.2 0 0")
-
-
-def assert_fsx_ontap_correctly_mounted(remote_command_executor, mount_dir, volume_id):
-    logging.info("Testing fsx Ontap is correctly mounted on the head node")
-    result = remote_command_executor.run_remote_command("df -h -t nfs4")
-    fsx_client = boto3.client("fsx")
-    volume = fsx_client.describe_volumes(VolumeIds=[volume_id]).get("Volumes")[0]
-    junction_path = volume["OntapConfiguration"]["JunctionPath"]
-    storage_virtual_machine_id = volume["OntapConfiguration"]["StorageVirtualMachineId"]
-    dns_name = boto3.client("fsx").describe_storage_virtual_machines(
-        StorageVirtualMachineIds=[storage_virtual_machine_id]
-    )["StorageVirtualMachines"][0]["Endpoints"]["Nfs"]["DNSName"]
-    remote_path = f"{dns_name}:{junction_path}"
-    assert_that(result.stdout).matches(rf"{remote_path} .* {mount_dir}")
-    # example output: "svm-123456.fs-123456.fsx.us-west-2.amazonaws.com:/vol1 9.5G 448K 9.5G 1% /fsx_mount_dir1"""
-    check_fstab_file(remote_command_executor, f"{remote_path} {mount_dir} nfs defaults 0 0")
-
-
-def check_fstab_file(remote_command_executor, expected_entry):
-    result = remote_command_executor.run_remote_command("cat /etc/fstab")
-    assert_that(result.stdout).matches(expected_entry)
-
-
-def get_mount_name(fsx_fs_id, region):
-    logging.info("Getting MountName from DescribeFilesystem API.")
-    fsx = boto3.client("fsx", region_name=region)
-    return (
-        fsx.describe_file_systems(FileSystemIds=[fsx_fs_id])
-        .get("FileSystems")[0]
-        .get("LustreConfiguration")
-        .get("MountName")
-    )
-
-
-def get_file_system_type(fsx_id, region):
-    fsx = boto3.client("fsx", region_name=region)
-    if fsx_id.startswith("fs-"):
-        logging.info("Getting file system type from DescribeFilesystems API.")
-        return fsx.describe_file_systems(FileSystemIds=[fsx_id]).get("FileSystems")[0].get("FileSystemType")
-    else:
-        logging.info("Getting file system type from FSx DescribeVolumes API.")
-        return fsx.describe_volumes(VolumeIds=[fsx_id]).get("Volumes")[0].get("VolumeType")
-
-
-def get_fsx_ids(cluster, region):
-    return utils.retrieve_cfn_outputs(cluster.cfn_name, region).get("FSXIds").split(",")
-
-
 def _get_storage_type(fsx):
     logging.info("Getting StorageType from DescribeFilesystem API.")
     return fsx.get("FileSystems")[0].get("StorageType")
@@ -561,40 +363,6 @@ def _get_deployment_type(fsx):
 def _test_deployment_type(deployment_type, fsx):
     logging.info("Test fsx deployment type")
     assert_that(_get_deployment_type(fsx)).is_equal_to(deployment_type)
-
-
-def _test_import_path(remote_command_executor, mount_dir):
-    logging.info("Testing fsx lustre import path")
-    result = remote_command_executor.run_remote_command("cat {mount_dir}/s3_test_file".format(mount_dir=mount_dir))
-    assert_that(result.stdout).is_equal_to("Downloaded by FSx Lustre")
-
-
-def assert_fsx_correctly_shared(scheduler_commands, remote_command_executor, mount_dir):
-    logging.info("Testing fsx correctly mounted on compute nodes")
-    remote_command_executor.run_remote_command("touch {mount_dir}/test_file".format(mount_dir=mount_dir))
-    job_command = "cat {mount_dir}/test_file && touch {mount_dir}/compute_output".format(mount_dir=mount_dir)
-    result = scheduler_commands.submit_command(job_command)
-    job_id = scheduler_commands.assert_job_submitted(result.stdout)
-    scheduler_commands.wait_job_completed(job_id)
-    scheduler_commands.assert_job_succeeded(job_id)
-    remote_command_executor.run_remote_command("cat {mount_dir}/compute_output".format(mount_dir=mount_dir))
-
-
-def _test_export_path(remote_command_executor, mount_dir, bucket_name, region):
-    logging.info("Testing fsx lustre export path")
-    remote_command_executor.run_remote_command(
-        "echo 'Exported by FSx Lustre' > {mount_dir}/file_to_export".format(mount_dir=mount_dir)
-    )
-    remote_command_executor.run_remote_command(
-        "sudo lfs hsm_archive {mount_dir}/file_to_export && sleep 5".format(mount_dir=mount_dir)
-    )
-    remote_command_executor.run_remote_command(
-        "sudo aws s3 cp --region {region} s3://{bucket_name}/export_dir/file_to_export ./file_to_export".format(
-            region=region, bucket_name=bucket_name
-        )
-    )
-    result = remote_command_executor.run_remote_command("cat ./file_to_export")
-    assert_that(result.stdout).is_equal_to("Exported by FSx Lustre")
 
 
 def _test_auto_import(auto_import_policy, remote_command_executor, mount_dir, bucket_name, region):
@@ -634,62 +402,6 @@ def _test_auto_import(auto_import_policy, remote_command_executor, mount_dir, bu
 def _assert_file_does_not_exist(remote_command_executor, filename, mount_dir):
     result = remote_command_executor.run_remote_command(f"ls {mount_dir}/")
     assert_that(result.stdout).does_not_contain(filename)
-
-
-@retry(
-    retry_on_result=lambda result: result.get("Lifecycle") in ["PENDING", "EXECUTING", "CANCELLING"],
-    wait_fixed=seconds(5),
-    stop_max_delay=minutes(7),
-)
-def poll_on_data_export(task, fsx):
-    logging.info(
-        "Data Export Task {task_id}: {status}".format(task_id=task.get("TaskId"), status=task.get("Lifecycle"))
-    )
-    return fsx.describe_data_repository_tasks(TaskIds=[task.get("TaskId")]).get("DataRepositoryTasks")[0]
-
-
-def _test_data_repository_task(remote_command_executor, mount_dir, bucket_name, fsx_fs_id, region):
-    logging.info("Testing fsx lustre data repository task")
-    file_contents = "Exported by FSx Lustre"
-    remote_command_executor.run_remote_command(
-        "echo '{file_contents}' > {mount_dir}/file_to_export".format(file_contents=file_contents, mount_dir=mount_dir)
-    )
-
-    # set file permissions
-    remote_command_executor.run_remote_command(
-        "sudo chmod 777 {mount_dir}/file_to_export && sudo chown 6666:6666 {mount_dir}/file_to_export".format(
-            mount_dir=mount_dir
-        )
-    )
-
-    fsx = boto3.client("fsx", region_name=region)
-    task = fsx.create_data_repository_task(
-        FileSystemId=fsx_fs_id, Type="EXPORT_TO_REPOSITORY", Paths=["file_to_export"], Report={"Enabled": False}
-    ).get("DataRepositoryTask")
-
-    task = poll_on_data_export(task, fsx)
-
-    assert_that(task.get("Lifecycle")).is_equal_to("SUCCEEDED")
-
-    remote_command_executor.run_remote_command(
-        "sudo aws s3 cp --region {region} s3://{bucket_name}/export_dir/file_to_export ./file_to_export".format(
-            region=region, bucket_name=bucket_name
-        )
-    )
-    result = remote_command_executor.run_remote_command("cat ./file_to_export")
-    assert_that(result.stdout).is_equal_to(file_contents)
-
-    # test s3 metadata
-    s3 = boto3.client("s3", region_name=region)
-    metadata = (
-        s3.head_object(Bucket=bucket_name, Key="export_dir/file_to_export").get("ResponseMetadata").get("HTTPHeaders")
-    )
-    file_owner = metadata.get("x-amz-meta-file-owner")
-    file_group = metadata.get("x-amz-meta-file-group")
-    file_permissions = metadata.get("x-amz-meta-file-permissions")
-    assert_that(file_owner).is_equal_to("6666")
-    assert_that(file_group).is_equal_to("6666")
-    assert_that(file_permissions).is_equal_to("0100777")
 
 
 def _test_storage_capacity(remote_command_executor, mount_dir, storage_capacity):
